requirements.txt
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
matplotlib==3.4.3
numpy==1.21.2
opencv-python-headless==4.4.0.46
pandas==1.3.3
Pillow==8.4.0
scipy==1.7.0
seaborn==0.11.1
streamlit==1.0.0
tensorflow-cpu==2.7.0


